requirements.txt
aiohttp==3.7.4.post0
anyio==3.4.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
discord.py==1.7.3
fastapi==0.70.1
h11==0.12.0
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pydantic==1.8.2
python-dotenv==0.19.2
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
yarl==1.7.2
